:construction: idk why the tests dont find fs.existsSync but i need to do something eles
